requirements.txt
Flask==0.10.1
Flask-Login==0.3.2
Flask-OAuth==0.12
Flask-RESTful==0.3.5
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.4
aniso8601==1.1.0
httplib2==0.9.2
itsdangerous==0.24
oauth2==1.9.0.post1
plaid-python==1.3.0
python-dateutil==2.5.0
pytz==2015.7
requests==2.3.0
six==1.10.0
validate-email==1.3
wsgiref==0.1.2
